requirements.txt
langchain-groq
langchain-community
python-dotenv
streamlit
ppprint
Openpyxl
pywin32